+++ examples/new_task.py
@@ -8,7 +8,7 @@ import sys
 
 async def new_task():
     try:
-        transport, protocol = await trio_amqp.connect('localhost', 5672)
+        protocol = await trio_amqp.connect('localhost', 5672)
     except trio_amqp.AmqpClosedConnection:
         print("closed connections")
         return
@@ -30,8 +30,7 @@ async def new_task():
     )
     print(" [x] Sent %r" % message,)
 
-    await protocol.close()
-    transport.close()
+    await protocol.aclose()
 
 
 trio.run(new_task)