+++ examples/receive.py
@@ -11,14 +11,17 @@ async def callback(channel, body, envelope, properties):
     print(" [x] Received %r" % body)
 
 async def receive():
-    transport, protocol = await trio_amqp.connect()
+    protocol = await trio_amqp.connect()
     channel = await protocol.channel()
 
     await channel.queue_declare(queue_name='hello')
 
     await channel.basic_consume(callback, queue_name='hello')
 
-    await trio.sleep_forever()
+    try:
+        await trio.sleep_forever()
+    finally:
+        await protocol.aclose()
 
 
 trio.run(receive)