+++ examples/receive_log_topic.py
@@ -19,7 +19,7 @@ async def callback(channel, body, envelope, properties):
 
 async def receive_log():
     try:
-        transport, protocol = await trio_amqp.connect('localhost', 5672)
+        protocol = await trio_amqp.connect('localhost', 5672)
     except trio_amqp.AmqpClosedConnection:
         print("closed connections")
         return
@@ -47,6 +47,9 @@ async def receive_log():
     print(' [*] Waiting for logs. To exit press CTRL+C')
 
     await channel.basic_consume(callback, queue_name=queue_name)
-    await trio.sleep_forever()
+    try:
+        await trio.sleep_forever()
+    finally:
+        await protocol.aclose()
 
 trio.run(receive_log)